refactor: 패키지 이동 및 클래스 이름 수정
