+++ src/main/java/lotto/constants/view/InputViewMessage.java
@@ -1,6 +1,6 @@
-package lotto.constants;
+package lotto.constants.view;
 
-public final class InputViewConstant {
+public final class InputViewMessage {
 
     public static final String ENTER_PAYMENT = "구입금액을 입력해 주세요.";
     public static final String ENTER_WINNING_NUMBERS = "당첨 번호를 입력해 주세요.";
