+++ src/main/java/lotto/view/InputView.java
@@ -1,7 +1,7 @@
 package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import lotto.constants.InputViewConstant;
+import lotto.constants.view.InputViewMessage;
 import lotto.model.draw_numbers.builder.DrawNumbersBuilder;
 import lotto.utils.PaymentValidator;
 
@@ -10,20 +10,20 @@ public class InputView {
     private final PaymentValidator paymentValidator = new PaymentValidator();
 
     public long getPayment() {
-        System.out.println(InputViewConstant.ENTER_PAYMENT);
+        System.out.println(InputViewMessage.ENTER_PAYMENT);
         String payment = readLine();
         paymentValidator.validate(payment);
         return Long.parseLong(payment);
     }
 
     public DrawNumbersBuilder getWinningNumbers(DrawNumbersBuilder builder) {
-        System.out.println(InputViewConstant.ENTER_WINNING_NUMBERS);
+        System.out.println(InputViewMessage.ENTER_WINNING_NUMBERS);
         String winningNumbers = readLine();
         return builder.winningNumbers(winningNumbers);
     }
 
     public DrawNumbersBuilder getBonusNumber(DrawNumbersBuilder builder) {
-        System.out.println(InputViewConstant.ENTER_BONUS_NUMBER);
+        System.out.println(InputViewMessage.ENTER_BONUS_NUMBER);
         String bonusNumber = readLine();
         return builder.bonusNumber(bonusNumber);
     }
